* Standardise pgm_async_create error code to -1 + errno. * Update pgm_async_recv to gsize/gssize lengths. * Update examples to use gsize/gssize for send/recv. * Fix error handling of examples with pgm_transport_bind.

+++ openpgm/pgm/test/app.c
@@ -445,7 +445,7 @@ session_send (
 	}
 
 /* send message */
-        int e = pgm_transport_send (sess->transport, string, strlen(string) + 1, 0);
+        gssize e = pgm_transport_send (sess->transport, string, strlen(string) + 1, 0);
         if (e < 0) {
 		puts ("FAILED: pgm_transport_send()");
         }
